requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
pyjwt==1.7.1
django-cors-headers==3.5.0
pymongo==3.12.3
djongo==1.3.6
djangotoolbox==1.8.0